requirements.txt
Flask
psycopg2
bcrypt
requests
gunicorn
python-dotenv